% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_enumerate_inputs.R
\name{GE_enumerate_inputs}
\alias{GE_enumerate_inputs}
\title{GE_enumerate_inputs.R}
\usage{
GE_enumerate_inputs()
}
\description{
Call this function only to display the necessary inputs for GE_bias.
}
\examples{
GE_enumerate_inputs()
}
\keyword{inputs}

