% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_enumerate_inputs_old.R
\name{GE_enumerate_inputs_old}
\alias{GE_enumerate_inputs_old}
\title{GE_enumerate_inputs_old.R}
\usage{
GE_enumerate_inputs_old()
}
\value{
Nothing
}
\description{
Call this function to display the necessary inputs for GE_bias.  If you see a term like
mu_Gf, that means it is the scalar E[G*f(E)].  If you see a term like MU_GM, that means it
is the vector c(E[G*M_1], E[G*M_2], ..., E[G*M_q]) where M is of dimension q.  For terms in 
cov_mat_list like E[ZZ], these should be matrices where the (i,j) element is E[Z_i*Z_j].
}
\examples{
GE_enumerate_inputs_old()
}

