\name{GEint-package}
\alias{GEint-package}
\alias{GEint}
\docType{package}
\title{
Gene-Environment Interaction
}
\description{
Software to calculate bias of coefficients in misspecified interaction models.  Also 
perform better inference for GWEIS.
}
\details{
\tabular{ll}{
Package: \tab GEint\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2016-05-25\cr
License: \tab GPL-2\cr
}

Right now only works with scalar arguments.
}
\author{
Ryan Sun

Maintainer: Ryan Sun <ryansun@fas.harvard.edu>
}
\references{
Coming soon!
}

\keyword{ interaction }
\seealso{
<~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~>
}
\examples{
GE_enumerate_inputs()
}
