% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_GE_bias_normal_squaredmis.R
\name{scalar_GE_bias_normal_squaredmis}
\alias{scalar_GE_bias_normal_squaredmis}
\title{GE_bias_normal_squaredmis}
\usage{
scalar_GE_bias_normal_squaredmis(beta_vec, rho_vec, prob_G)
}
\arguments{
\item{beta_vec}{A vector of the effect sizes in the true model.
Use the order beta_0, beta_G, beta_E, beta_I, beta_Z, beta_M.}

\item{rho_vec}{A vector of the 10 pairwise covariances between the
covariates.  These should be in the order (1) cov_GE (2) cov_GZ (3) cov_EZ
(4) cov_GW (5) cov_EW (6) cov_ZW.}

\item{prob_G}{Probability that each allele is equal to 1.  Since each SNP has
two alleles, the expectation of G is 2*prob_G.}
}
\description{
Deprecated version as of Sep 27, 2016.  Designed to work only with scalar inputs for Z and W/M.
}
\details{
A function to calculate the bias in testing for GxE interaction.
Here we make the following assumptions:
(1) All covariates (besides E) are marginally normally distributed with 
unit variance.  In addition, E,Z,W have mean 0.
(2) G is generated by means of thresholding two independent normal RVs (jointly generated
with the other covariates) and is centered to have mean 0.
(3) The misspecification is of the form f(E)=h(E)=E^2, and M=W^2.
(4) All covariate terms exist and are scalars.
}
\examples{
scalar_GE_bias_normal_squaredmis( beta_vec=runif(n=6, min=0, max=1), rho_vec=rep(0.3,1), prob_G=0.3)
}
\keyword{bias}
\keyword{normal}
\keyword{squared_misspecification}

