% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_GE_nleqslv.R
\name{scalar_GE_nleqslv}
\alias{scalar_GE_nleqslv}
\title{GE_nleqslv}
\usage{
scalar_GE_nleqslv(beta_vec, mu_vec, cov_vec, MU_ZZ, MU_ZM, MU_ZW, MU_WM, MU_WW,
  HOM_vec)
}
\arguments{
\item{beta_vec}{A vector of the effect sizes in the true model.
Use the order beta_0, beta_G, beta_E, beta_I, beta_Z, beta_M.}

\item{mu_vec}{A vector of the means for f, h, M, W, Z, in that order.}

\item{cov_vec}{A vector of the covariances in the order
(mu_GG, mu_GE, mu_Gf, mu_Gh, MU_GZ, MU_GM, MU_GW, mu_EE,
mu_Ef, MU_EZ, MU_EM, MU_EW, MU_fZ, MU_fW)}

\item{MU_ZZ}{Matrix mean}

\item{MU_ZM}{Matrix mean}

\item{MU_ZW}{Matrix mean}

\item{MU_WM}{Matrix mean}

\item{MU_WW}{Matrix mean}

\item{HOM_vec}{A vector of the higher order moments in the order
(mu_GGE, mu_GGh, mu_GEE, mu_GEF, mu_GEh, MU_GEZ, MU_GEM, MU_GEW,
MU_GhW, MU_GhZ, mu_GGEE, mu_GGEf, mu_GGEh)}

\item{MU_WZ}{Matrix mean}
}
\description{
Deprecated version as of Sep 27, 2016.  Designed only to work with scalar Z and W/M.
}
\details{
Here we use the package nleqslv to get a numerical solution which
we can use to check our direct solution.
}
\examples{
GE_bias_results <- GE_bias_normal_squaredmis(runif(6), runif(3), runif(10))
GE_nleqslv(GE_bias_results$beta_vec, GE_bias_results$mu_vec, GE_bias_results$cov_vec,
GE_bias_results$MU_ZZ, GE_bias_results$MU_ZM, GE_bias_results$MU_ZW, GE_bias_results$MU_WZ,
GE_bias_results$MU_WM, GE_bias_results$MU_WW, GE_bias_results$HOM_vec)
}
\keyword{equation}
\keyword{nonlinear}

