% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_test_normal_calcs.R
\name{scalar_GE_normal_test}
\alias{scalar_GE_normal_test}
\title{test_normal_calcs}
\usage{
scalar_GE_normal_test(rho_vec, prob_G, test_threshold = 0.01)
}
\arguments{
\item{rho_vec}{A vector of the 10 pairwise covariances between the
covariates.  These should be in the order (1) cov_GE (2) cov_GZ (3) cov_EZ
(4) cov_GW (5) cov_EW (6) cov_ZW.}

\item{prob_G}{Probability that each allele is equal to 1.  Since each SNP has
two alleles, the expectation of G is 2*prob_G.}

\item{test_threshold}{How much margin for error on tests?}
}
\description{
Deprecated as of Sep 27, 2016.  Designed to work only with scalar Z and W/M.
}
\details{
Only a test function to ensure the quantities calculated in GE_bias_normal_squaredmis
are correct.
}

