% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_bias_normal_squaredmis.R
\name{GE_bias_normal_squaredmis}
\alias{GE_bias_normal_squaredmis}
\title{GE_bias_normal_squaredmis}
\usage{
GE_bias_normal_squaredmis(beta_vec, means_vec, rho_vec)
}
\arguments{
\item{beta_vec}{A vector of the effect sizes in the true model.
Use the order beta_0, beta_G, beta_E, beta_I, beta_Z, beta_M.}

\item{means_vec}{A vector of the means for M, W, Z, in that order.}

\item{rho_vec}{A vector of the 10 pairwise correlations between the
covariates.  Suppose that the order of the rows/columns is G,E,Z,M,W.
Then the correlation in row i, column j should be the (i-1)*(i-2)/2 + j
element in the vector.}
}
\description{
A function to calculate the bias in testing for GxE interaction.
Here we make the following assumptions:
(1) All covariates are marginally normally distributed with 
unit variance.  In addition, G and E have mean 0.
(2) The misspecification is of the form f(E)=h(E)=E^2
(3) All covariate terms exist and are scalars
}
\examples{
GE_bias_normal_squaredmis( runif(6), runif(3), runif(10) )
}
\keyword{bias}
\keyword{normal}
\keyword{squared_misspecification}

