% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_enumerate_inputs.R
\name{GE_enumerate_inputs}
\alias{GE_enumerate_inputs}
\title{GE_enumerate_inputs.R}
\usage{
GE_enumerate_inputs()
}
\value{
Nothing
}
\description{
Call this function to display the necessary inputs for GE_bias_set. For terms in 
cov_mat_list like E[ZZ], these should be matrices where the (i,j) element is E[Z_i*Z_j].
}
\examples{
GE_enumerate_inputs()
}
