% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_scoreeq_sim.R
\name{GE_scoreeq_sim}
\alias{GE_scoreeq_sim}
\title{GE_scoreeq_sim.R}
\usage{
GE_scoreeq_sim(num_sims = 5000, num_sub = 2000, beta_vec, mu_vec, rho_vec)
}
\arguments{
\item{num_sims}{The number of repetitions to run, default is 5000}

\item{num_sub}{The number of subjects to simulate, default is 2000}

\item{beta_vec}{A vector of the effect sizes in the true model.
Use the order beta_0, beta_G, beta_E, beta_I, beta_Z, beta_M.}

\item{mu_vec}{A vector of the means for M, W, Z, in that order.}

\item{rho_vec}{A vector of the covariances in the order:
(rho_GE, rho_GZ, rho_GM, rho_GW, rho_EZ, rho_EM, rho_EW, rho_ZM, rho_ZW, rho_WM)}
}
\description{
Here we perform simulation to verify that we have solved for
the correct alpha values.
}
\examples{
library(mvtnorm)
temp_sig <- matrix(data=0.5, nrow=5, ncol=5)
diag(temp_sig) <- 1
sig_mat <- cor( rmvnorm(n=10, sigma=temp_sig) )
rho_vec <- sig_mat[upper.tri(sig_mat)]
mu_vec <- runif(3)
beta_vec <- runif(6)
GE_bias_results <- GE_bias_normal_squaredmis(beta_vec, mu_vec, rho_vec)
GE_nleqslv(GE_bias_results$beta_vec, GE_bias_results$mu_vec, GE_bias_results$cov_vec,
GE_bias_results$MU_ZZ, GE_bias_results$MU_ZM, GE_bias_results$MU_ZW, GE_bias_results$MU_WZ,
GE_bias_results$MU_WM, GE_bias_results$MU_WW, GE_bias_results$HOM_vec)
GE_scoreeq_sim(beta_vec=beta_vec, mu_vec=mu_vec, rho_vec=rho_vec)
}
\keyword{simulation}

